+++ src/components/input/input.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import * as lifecycle from './lifecycle/lifecycle'
 import * as handlers from './handlers/handlers'
 import { decorator } from './decorator/decorator'
 
@@ -7,9 +6,9 @@ export default class ManagedInput extends React.Component {
   constructor(props) {
     super(props)
     this.input = React.createRef()
-    this.controlled = this.props.manager('controlled')
-    this.id = this.props.manager('id')
-    this.classNames = this.props.manager('classNames')
+    this.controlled = this.props.controlled
+    this.id = this.props.id
+    this.classNames = this.props.classNames || {}
     this.inputClass = this.classNames.input || 'input'
     this.labelClass = this.classNames.label || 'label'
     this.valueClass = this.classNames.value || 'value'
@@ -18,9 +17,6 @@ export default class ManagedInput extends React.Component {
     this.hoverClass = this.classNames.hover || 'hover'
     this.errorClass = this.classNames.error || 'error'
     this.errorContainerClass = this.classNames.error || 'error-container'
-    this.componentDidMount = lifecycle.componentDidMount
-    this.shouldComponentUpdate = lifecycle.shouldComponentUpdate
-    this.componentDidUpdate = lifecycle.componentDidUpdate
     this.onMouseOverHandler = handlers.onMouseOverHandler.bind(this)
     this.onMouseLeaveHandler = handlers.onMouseLeaveHandler.bind(this)
     this.onFocusHandler = handlers.onFocusHandler.bind(this)
@@ -28,24 +24,12 @@ export default class ManagedInput extends React.Component {
     this.onChangeHandler = handlers.onChangeHandler.bind(this)
     this.onClickHandler = handlers.onClickHandler.bind(this)
     this.decorator = decorator.bind(this)
-    this.stateKeysToDerive = ['display', 'disabled', 'value', 'label', 'error']
-    this.stateKeysToUpdate = this.stateKeysToDerive.concat([ 'mouseOver', 'focus', 'decorator'])
     this.state={
-      stateKeysToDerive: this.stateKeysToDerive,
-      display: this.props.manager('display'),
-      disabled: this.props.manager('disabled'),
-      value: this.props.manager('value'),
-      label: this.props.manager('label'),
-      error: this.props.manager('error'),
       mouseOver: false,
       autofill: true,
       focus: false,
       decorator: ''
     }
-  }
-
-  static getDerivedStateFromProps(nextProps, prevstate) {
-    return lifecycle.getDerivedStateFromProps(nextProps, prevstate)
   }
 
   isMobile() {