+++ src/components/pagination/handlers/handlers.js
@@ -1,41 +1,52 @@
 export function onMouseOverHandler(button) {
-  let mouseOver = {...this.state.mouseOver}
+  const { onMouseOver } = this.props
+  let mouseOver = { ...this.state.mouseOver }
   mouseOver[button] = true
-  this.setState({ mouseOver: mouseOver })
-  this.props.manager('onMouseOver', button)
+  this.setState({
+    mouseOver,
+    decorator: this.decorator({ mouseOver })
+  })
+  onMouseOver && onMouseOver(button)
 }
 
 export function onMouseLeaveHandler(button) {
-  let mouseOver = {...this.state.mouseOver}
+  const { onMouseLeave } = this.props
+  let mouseOver = { ...this.state.mouseOver }
   mouseOver[button] = false
-  this.setState({ mouseOver: mouseOver })
-  this.props.manager('onMouseOver', button)
+  this.setState({
+    mouseOver,
+    decorator: this.decorator({ mouseOver })
+  })
+  onMouseLeave && onMouseLeave(button)
 }
 
 export function onFocusHandler(button) {
-  let focus = {...this.state.focus}
+  const { onFocus } = this.props
+  let focus = { ...this.state.focus }
   focus[button] = true
-  this.setState({ focus: focus })
-  this.props.manager('onFocus', button)
+  this.setState({
+    focus,
+    decorator: this.decorator({ focus })
+  })
+  onFocus && onFocus(button)
 }
 
 export function onBlurHandler(button) {
-  let focus = {...this.state.focus}
+  const { onBlur } = this.props
+  let focus = { ...this.state.focus }
   focus[button] = false
-  this.setState({ focus: focus })
-  this.props.manager('onFocus', button)
+  this.setState({
+    focus,
+    decorator: this.decorator({ focus })
+  })
+  onBlur && onBlur(button)
 }
 
 export function onClickHandler(button) {
-  if (button === 0 || button === this.state.buttons.length - 1) {
-    let current
-    if (button === 0) {
-      current = this.state.current - 1
-    } else {
-      current = this.state.current + 1
-    }
-    this.props.manager('onClick', current)
-    return
-  }
-  this.props.manager('onClick', button)
+  const { onClick } = this.props
+  let current
+  if (button === 0) current = this.props.current - 1
+  else if (button === this.state.buttons.length - 1) current = this.props.current + 1
+  else current = button
+  onClick(current)
 }