+++ src/components/radio/radio.js
@@ -7,14 +7,10 @@ export default class ManagedRadio extends ManagedInput {
   constructor(props) {
     super(props)
     this.input = []
-    this.name = this.props.manager('name')
-    this.radios = this.props.manager('radios')
     this.containerClass = this.classNames.container || 'managed-input radio'
     this.radioContainerClass = this.classNames.radioContainerClass || 'container'
     this.checkedClass = this.classNames.checked || 'checked'
     this.tickClass = this.classNames.tick || 'tick'
-    this.stateKeysToDerive = [...this.stateKeysToDerive, 'checked']
-    this.stateKeysToUpdate = [...this.stateKeysToUpdate, 'checked']
     this.onMouseOverHandler = handlers.onMouseOverHandler.bind(this)
     this.onMouseLeaveHandler = handlers.onMouseLeaveHandler.bind(this)
     this.onFocusHandler = handlers.onFocusHandler.bind(this)
@@ -22,45 +18,43 @@ export default class ManagedRadio extends ManagedInput {
     this.onKeyDownHandler = handlers.onKeyDownHandler.bind(this)
     this.onClickHandler = handlers.onClickHandler.bind(this)
     this.decorator = decorator.bind(this)
-    this.state = {
-      ...this.state,
-      stateKeysToDerive: this.stateKeysToDerive,
-      checked: this.props.manager('checked')
-    }
   }
 
   render() {
+    const { display, name, radios, checked, children } = this.props
+    const { decorator } = this.state
+
     return (
-      <div style={{ display: this.state.display }}>
+      <div style={{ display }}>
         {
-        this.radios.map((radio, index) => (
+        radios.map((radio, index) => (
         <div
-        className={this.containerClass + this.state.decorator[index]}
+        className={`${this.containerClass}${radio.value === checked ? ' checked' : ''}${decorator[index] ? decorator[index] : ''}`}
         key={radio.key || `radio-${this.name}-${index}`}
         onClick={() => this.onClickHandler(radio.value, index)}>
           <div
-          className={this.inputClass + this.state.decorator[index]}
+          className={`${this.inputClass}${radio.value === checked ? ' checked' : ''}${decorator[index] ? decorator[index] : ''}`}
           onMouseOver={(e) => this.onMouseOverHandler(e, index)}
           onMouseLeave={(e) => this.onMouseLeaveHandler(e, index)}>
             {
-            radio.value === this.state.checked &&
-            <div className={this.tickClass + this.state.decorator[index]} />
+            radio.value === checked &&
+            <div className={`${this.tickClass}${decorator[index] ? decorator[index] : ''}`} />
             }
           </div>
           <input
-          name={this.name}
+          name={name}
           ref={this.input[index] = React.createRef()}
           id={radio.id}
           type='radio'
           style={{ position: 'absolute', left: -10000 }}
           value={radio.value}
-          checked={radio.value === this.state.checked}
+          checked={radio.value === checked}
           onFocus={(e) => this.onFocusHandler(e, index)}
           onBlur={(e) => this.onBlurHandler(e, index)}
           onKeyDown={this.onKeyDownHandler}
           onChange={(e) => this.onChangeHandler(e)} />
           <label
-          className={this.labelClass + this.state.decorator[index]}
+          className={`${this.labelClass}${radio.value === checked ? ' checked' : ''}${decorator[index] ? decorator[index] : ''}`}
           onMouseOver={(e) => this.onMouseOverHandler(e, index)}
           onMouseLeave={(e) => this.onMouseLeaveHandler(e, index)}>
             {radio.label}
@@ -70,7 +64,7 @@ export default class ManagedRadio extends ManagedInput {
         ))
         }
         <div style={{ clear: 'both' }} />
-        {this.props.manager('children')}
+        {children}
       </div>
     )
   }