+++ src/components/text/text.js
@@ -5,40 +5,44 @@ import * as handlers from './handlers/handlers'
 export default class ManagedText extends ManagedInput {
   constructor(props) {
     super(props)
-    this.type = this.props.manager('type')
-    this.autoComplete = this.props.manager('autoComplete')
-    this.match = this.props.manager('match')
+    this.input = React.createRef()
+    this.type = this.props.type
+    this.autoComplete = this.props.autoComplete
+    this.match = this.props.match
     this.containerClass = this.classNames.container || 'managed-input text'
     this.onKeyDownHandler = handlers.onKeyDownHandler.bind(this)
     this.onChangeHandler = handlers.onChangeHandler.bind(this)
   }
 
   render() {
+    const { display, disabled, value, label, children, error } = this.props
+    const decorator = `${this.state.decorator}${error ? ' error' : ''}${this.input.current && this.input.current.value ? ' value' : ''}`
+
     return (
       <div
-      style={{ display: this.state.display }}
-      className={this.containerClass + this.state.decorator}
+      style={{ display }}
+      className={this.containerClass + decorator}
       onMouseOver={this.onMouseOverHandler}
       onMouseLeave={this.onMouseLeaveHandler}
       onClick={this.onClickHandler}>
-        <div className={this.labelClass + this.state.decorator}>
-          {this.state.label}
+        <div className={this.labelClass + decorator}>
+          {label}
         </div>
         <input
         ref={this.input}
         type={this.type}
-        disabled = {this.state.disabled}
+        disabled = {disabled}
         id={this.id}
-        className={this.inputClass + this.state.decorator}
+        className={`${this.inputClass}${decorator}`}
         autoComplete = {this.autoComplete}
-        value={this.controlled ? this.state.value : undefined}
+        value={this.controlled ? value : undefined}
         onFocus={this.onFocusHandler}
         onKeyDown={this.onKeyDownHandler}
         onBlur={this.onBlurHandler}
         onChange={this.onChangeHandler} />
-        {this.props.manager('children')}
-        <div className={this.errorContainerClass + this.state.decorator}>
-          {this.state.error}
+        {children}
+        <div className={this.errorContainerClass + decorator}>
+          {error}
         </div>
       </div>
     )