+++ src/components/textarea/handlers/handlers.js
@@ -1,9 +1,9 @@
 export function onChangeHandler(e) {
+  const { counterLimit, onChange } = this.props
   const valueLength = e.target.value.length
-  if (valueLength <= this.state.counterLimit) {
+  if (valueLength <= counterLimit) {
     this.setState({ counter: valueLength })
-    this.props.manager('onChange', e.target.value)
-    this.forceUpdate()
+    onChange(e.target.value)
   } else {
     this.input.current.value = this.input.current.value.slice(0, -1)
   }