+++ src/index.js
@@ -6,7 +6,6 @@ import ManagedCheckbox from './components/checkbox/checkbox'
 import ManagedMultipleCheckbox from './components/multiple-checkbox/multiple-checkbox'
 import ManagedButton from './components/button/button'
 import ManagedPagination from './components/pagination/pagination'
-import managerAgent from './manager-agent/manager-agent'
 
 export { ManagedText,
          ManagedSelect,
@@ -15,5 +14,4 @@ export { ManagedText,
          ManagedButton,
          ManagedCheckbox,
          ManagedMultipleCheckbox,
-         ManagedPagination,
-         managerAgent }
+         ManagedPagination }